+++ src_new/triggers/levi_ataxia/for_levi/leviticus/locate/003_Party_Locate.lua
@@ -38,4 +38,4 @@ patterns:
 
 locateon = true
 locateperson = matches[2]
-send("cq all;scry for " .. matches[3])
+send("cq all;farsee " .. matches[3])