+++ src_new/triggers/levi_ataxia/for_levi/leviticus/locate/004_Party_Locate_Scry.lua
@@ -10,7 +10,7 @@ hierarchy:
 - MINE ALL MINE
 - Locate
 attributes:
-  isActive: 'yes'
+  isActive: 'no'
   isFolder: 'no'
   isTempTrigger: 'no'
   isMultiline: 'yes'